Add status tracking and audit logging for category creation, deletion, and updates; refactor deleteCategory route to use POST method

+++ backend/controllers/categoryVendorController.js
@@ -1,7 +1,7 @@
 import asyncHandler from "express-async-handler";
 import categoryModel from "../models/categoryModel.js";
 import supplierModel from "../models/supplierModel.js";
-
+import AuditSupplierLog from "../models/auditSupplierModel.js";
 const getAllCategory = asyncHandler(async (req, res) => {
   const { userId } = req.body;
 
@@ -22,15 +22,35 @@ const getAllCategory = asyncHandler(async (req, res) => {
 });
 
 const createCategory = asyncHandler(async (req, res) => {
-  const { userId } = req.body;
+  const { userId, status } = req.body;
   const { category_name } = req.body;
-  
+
+  const existSupplier = await supplierModel.findById(userId);
+  if (!existSupplier) {
+    return res
+      .status(400)
+      .json({ success: false, message: "Supplier Id not found!" });
+  }
   const newCategory = new categoryModel({
     category_name,
     supplier: userId,
   });
 
   await newCategory.save();
+
+  const newAuditLog = new AuditSupplierLog({
+    eventTypes: status,
+    entityType: "categories",
+    entityId: newCategory._id,
+    changes: {
+      oldValue: null,
+      newValue: newCategory,
+    },
+    performeBy: userId,
+    role: existSupplier.role,
+  });
+
+  await newAuditLog.save();
 
   res.status(201).json({ success: true, message: "Successfully created" });
 });
@@ -51,6 +71,23 @@ const getSpecificCategory = asyncHandler(async (req, res) => {
 
 const deleteCategory = asyncHandler(async (req, res) => {
   const { id } = req.params;
+  const { userId, status } = req.body;
+
+  const existSupplier = await supplierModel.findById(userId);
+
+  if (!existSupplier) {
+    return res
+      .status(400)
+      .json({ success: false, message: "Supplier Id not found!" });
+  }
+
+  const existCategory = await categoryModel.findById(id);
+
+  if (!existCategory) {
+    return res
+      .status(400)
+      .json({ success: false, message: "Categroy Id not found!" });
+  }
   const deletedCategory = await categoryModel.findByIdAndDelete(id);
 
   if (!deletedCategory) {
@@ -59,12 +96,26 @@ const deleteCategory = asyncHandler(async (req, res) => {
       .json({ success: false, message: "Category not found" });
   }
 
+  const newAuditLog = new AuditSupplierLog({
+    eventTypes: status,
+    entityType: "categories",
+    entityId: id,
+    changes: {
+      oldValue: existCategory,
+      newValue: deletedCategory,
+    },
+    performeBy: userId,
+    role: existSupplier.role,
+  });
+
+  await newAuditLog.save();
+
   res.status(200).json({ success: true, message: "Category Deleted" });
 });
 
 const updateCategory = asyncHandler(async (req, res) => {
   const { id } = req.params;
-
+  const { category_code, category_name } = req.body;
   const { userId } = req.body;
   const userExist = await supplierModel.findById(userId);
   if (!userExist) {
@@ -72,7 +123,22 @@ const updateCategory = asyncHandler(async (req, res) => {
       .status(400)
       .json({ success: false, message: "Supplier not found" });
   }
-  const { category_code, category_name } = req.body;
+
+  const existSupplier = await supplierModel.findById(userId);
+
+  if (!existSupplier) {
+    return res
+      .status(400)
+      .json({ success: false, message: "Supplier Id not found!" });
+  }
+
+  const existCategory = await categoryModel.findById(id);
+  if (!existCategory) {
+    return res
+      .status(400)
+      .json({ success: false, message: "Supplier Id not found!" });
+  }
+
   const updatedCategory = await categoryModel.findByIdAndUpdate(
     id,
     {
@@ -87,6 +153,19 @@ const updateCategory = asyncHandler(async (req, res) => {
       .status(400)
       .json({ success: false, message: "Category not found" });
   }
+  const newAuditLog = new AuditSupplierLog({
+    eventTypes: "Update",
+    entityType: "categories",
+    entityId: id,
+    changes: {
+      oldValue: existCategory,
+      newValue: updatedCategory,
+    },
+    performeBy: userId,
+    role: existSupplier.role,
+  });
+
+  await newAuditLog.save();
 
   res.status(200).json({ success: true, message: "Updated Successfully" });
 });
